Fix the server unit tests
The 'The Station model does not allow duplicate station numbers' unit
test used to fail because of a failing promise (that was expected to
fail). Because the promise was very hard to deal with, instead of that I
updated the ES version to ES2015 and used async/await to simplify
the test.

Unfortunately this also led to the Bluebird Promise being incompatible
with the native Promise, which then led to changes to other places.
Actually, `Promise.resolve` does the trick just fine, but I found that
out after I had already changed many functions to async ones.

NOTE: The cost of the readability of async functions is that they
are somewhat slower. Then again, promises themselves are
2-4 times slower than callbacks.